Specification e Validation - Categoria II

+++ DDDDemo.Dominio/Servicos/Base/ServiceBase.cs
@@ -1,5 +1,7 @@
 ﻿using DDDDemo.Dominio.Interfaces.Repositorio.Base;
 using DDDDemo.Dominio.Interfaces.Servico.Base;
+using DDDDemo.Dominio.Interfaces.Validation;
+using DomainValidation.Validation;
 using System;
 using System.Collections.Generic;
 
@@ -8,15 +10,29 @@ namespace DDDDemo.Dominio.Servicos.Base
     public class ServiceBase<TEntity> : IDisposable, IServiceBase<TEntity> where TEntity : class
     {
         private readonly IRepositoryBase<TEntity> _repository;
+        private readonly ValidationResult _validationResult;
+
+        protected ValidationResult ValidationResult
+        {
+            get { return _validationResult; }
+        }
 
         public ServiceBase(IRepositoryBase<TEntity> repository)
         {
             _repository = repository;
+            _validationResult = new ValidationResult();
         }
 
-        public void Add(TEntity obj)
+        public virtual ValidationResult Add(TEntity obj)
         {
+            if (!ValidationResult.IsValid)
+                return ValidationResult;
+
+            if (obj is ISelfValidation selfValidationEntity && !selfValidationEntity.IsValid)
+                return selfValidationEntity.ValidationResult;
+
             _repository.Add(obj);
+            return _validationResult;
         }
 
         public void Dispose()
@@ -34,14 +50,25 @@ namespace DDDDemo.Dominio.Servicos.Base
             return _repository.GetById(id);
         }
 
-        public void Remove(TEntity obj)
+        public ValidationResult Remove(TEntity obj)
         {
+            if (!ValidationResult.IsValid)
+                return ValidationResult;
+
             _repository.Remove(obj);
+            return _validationResult;
         }
 
-        public void Update(TEntity obj)
+        public ValidationResult Update(TEntity obj)
         {
+            if (!ValidationResult.IsValid)
+                return ValidationResult;
+
+            if (obj is ISelfValidation selfValidationEntity && !selfValidationEntity.IsValid)
+                return selfValidationEntity.ValidationResult;
+
             _repository.Update(obj);
+            return _validationResult;
         }
     }
 }
